Moved catalog and rdd to geopyspark.geotrellis
Removed constants and rdd from geopyspark
